Add files via upload
For this project, we created a DBMS within Microsoft Access for Apple Inc. We designed relational tables and their corresponding relationships relating to Apple's brick-and-mortar stores (such as Customer, Store, Distributor, Supplier) in order to generate an efficient way for Apple to track purchases and forecast demand for their products that are sold in physical stores. Using SQL, we also created queries in order to generate forms and reports.